Implement BlackMagic controls and rename to MultiRecorder
