+++ README.md
@@ -1,13 +1,14 @@
-# OBS Controller
+# MultiRecorder
 
-A simple Python based utility for controlling and monitoring multiple OBS instances over WebSocket.
-
-![image](https://github.com/EvanPeacock/OBS-Controller/assets/36444106/2f36491c-d255-403a-a540-b9748496f40a)
+A simple Python based utility for controlling and monitoring multiple OBS instances and BlackMagic HyperDeck products over a network connection.
 
 
-For each connected OBS instance, OBS Controller:
-* Displays an inital screenshot
+For each connected OBS instance, MultiRecorder:
+* Displays an inital screenshot (if enabled via command line arguments)
 * Displays current framerate, resolution, and recording time
 * Allows for stopping, starting, pausing, and unpausing of recording
 
-Support for similarly remotely controlling Blackmagic HyperDeck products is planned.
+
+For each BlackMagic HyperDeck product, MultiRecorder:
+* Displays current framerate, resolution, video interface, codec, and recording time
+* Allows for stopping and starting, of recording
