--- controller/config.yaml
@@ -1,7 +1,0 @@
-obs_connections:
-  - name: "Local OBS"
-    host: "localhost"
-    port: 4455
-# - name: "Some other connection"
-#   host: "192.168.1.1"
-#   port: 4455
