+++ multi_recorder/assets/icon.ico
