+++ multi_recorder/requirements.txt
@@ -1,0 +1,5 @@
+pyaml
+dearpygui
+obsws_python
+pillow
+requests
