+++ standalone_packager/excluded_file_list.txt
@@ -5,4 +5,3 @@ __pycache__
 .pyd\
 .gz\
 .zip\
-vc140.pdb
