+++ standalone_packager/runner_files/Start MultiRecorder with Previews.bat
@@ -1,0 +1,16 @@
+@echo off
+
+set "CONFIG_FILE=config.yaml"
+
+echo Using CONFIG_FILE = %CONFIG_FILE%
+echo.
+echo.
+echo Starting MultiRecorder...
+echo.
+echo.
+
+cd multi_recorder
+MultiRecorder.exe --config-file "%CONFIG_FILE%" -p True
+
+:end
+timeout /t 30
